IntelligentFaceMatch/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.12",
    "flask-socketio>=5.5.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "gunicorn>=23.0.0",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "prometheus-client>=0.21.1",
    "protobuf>=5.29.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "redis>=5.2.1",
    "scikit-image>=0.25.2",
    "sqlalchemy>=2.0.40",
    "starlette>=0.46.1",
    "uvicorn>=0.34.0",
]
